fixing calculator
