+++ src/stack_calculator/calculator_types.h
@@ -28,6 +28,7 @@ typedef enum
 typedef enum
 {
   TOKEN_OK,
+  TOKEN_INVALID_ALPHA,
   TOKEN_INVALID_UNDEFINED,
   TOKEN_INVALID_NUMBER,
   TOKEN_INVALID_OPERATOR,
