+++ src/stack_calculator/debug.h
@@ -1,0 +1,4 @@
+#include "stack.h"
+
+void print_stack(const Stack *stack);
+void print_source(const char *source);
