+++ src/stack_calculator/stack.c
@@ -91,20 +91,3 @@ bool stack_is_empty(const Stack* stack) {
 
   return stack->top == 0;
 }
-
-void print_stack(const Stack* stack) {
-  if (stack->top > INT_MAX) {
-    // Handle error case
-    printf("Stack too large for int conversion\n");
-    return;
-  }
-
-  printf("------\n");
-  printf("Stack:\n");
-  printf("------\n");
-
-  for (int i = 0; i < (int)stack->top; i++) {
-    printf("Index: %d, Value: %f\n", i, stack->values[i]);
-  }
-  printf("\n");
-}
